Show VC verification status (#61)
* feat: show verification result in step2

* feat: download, refresh and delete VCs

Works for step 2 now, step 3 is broken, will be fixed in next commit

* fix: updated step 3 with new feature

* refactor: move link viewing to VC comp

* fix: downloadVC should check if file is available

When saving a file to the solid pod, it isn't immediately available. For this reason, wait a bit after saving the VC to the pod until it is ready

* feat: add loading icon when waiting for VC

Sometimes takes a couple of seconds to properly persist in Solid Pod. Keep polling until the saved file has become available to load

* refactor: some naming in VC saving