feat: login with own pod
People interested in the demo can now log in using their own Pod. Since we only have Vera and Koos data available in the backend, send a request to the backend to add dummy data for the new webId. In Step 1, get the data from the Solid Pod and use the publicly available data to bootstrap some demo data